refactor(encoding): consolidate windows-1252 codec into shared module
Move decode logic out of parser/io.rs and colocate it with the new encode
logic in a crate-level encoding module, so the two tables stay exact inverses.
Adds a round-trip test over all 256 bytes.
